backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# AI & RAG
openai==1.10.0
anthropic==0.18.1
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16

# Scraping (for future use)
playwright==1.41.0
firecrawl-py==0.0.5

# Utils
python-multipart==0.0.6
httpx==0.26.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
